+++ src/server/game/Entities/Player/Player.h
@@ -980,7 +980,7 @@ class TradeData
         Item*  GetSpellCastItem() const;
         bool HasSpellCastItem() const { return m_spellCastItem != 0; }
 
-        uint32 GetMoney() const { return m_money; }
+        uint64 GetMoney() const { return m_money; }
         void SetMoney(uint32 money);
 
         bool IsAccepted() const { return m_accepted; }
@@ -1528,9 +1528,9 @@ class Player : public Unit, public GridObject<Player>
         void setRegenTimerCount(uint32 time) {m_regenTimerCount = time;}
         void setWeaponChangeTimer(uint32 time) {m_weaponChangeTimer = time;}
 
-        uint32 GetMoney() const { return GetUInt32Value (PLAYER_FIELD_COINAGE); }
+        uint64 GetMoney() const { return GetUInt64Value (PLAYER_FIELD_COINAGE); }
         void ModifyMoney(int32 d);
-        bool HasEnoughMoney(uint32 amount) const { return (GetMoney() >= amount); }
+        bool HasEnoughMoney(uint64 amount) const { return (GetMoney() >= amount); }
         bool HasEnoughMoney(int32 amount) const
         {
             if (amount > 0)
