+++ src/server/game/Mails/Mail.h
@@ -183,7 +183,7 @@ struct Mail
     time_t expire_time;
     time_t deliver_time;
     uint32 money;
-    uint32 COD;
+    uint64 COD;
     uint32 checked;
     MailState state;
